Rename scope/ to symbol/
