+++ framework/CMakeLists.txt
@@ -44,7 +44,7 @@ set(SOURCES
     "src/type/StructType.cpp"
     "src/type/ArrayType.cpp"
 
-    "src/scope/Scope.cpp"
+    "src/symbol/Scope.cpp"
 )
 
 set(HEADERS
@@ -87,7 +87,7 @@ set(HEADERS
     "include/type/StructType.h"
     "include/type/ArrayType.h"
 
-    "include/scope/Scope.h"
+    "include/symbol/Scope.h"
 )
 
 source_group(TREE ${PROJECT_SOURCE_DIR} FILES ${SOURCES} ${HEADERS})
