+++ framework/include/symbol/Scope.h
